feat: ajustar permissões de autenticação e melhorar verificação de regras de usuário

+++ app/Http/Controllers/Admin/ActivityController.php
@@ -20,7 +20,7 @@ class ActivityController extends Controller
      */
     public function index(Request $request): Response
     {
-        $this->authorize('activities.showAny', Activity::class);
+        $this->authorize('activities.viewAny', Activity::class);
 
         return Inertia::render('Auth/Activity/Index', array_merge(ModelsActivity::search($request), [
             'can' => [
@@ -57,6 +57,8 @@ class ActivityController extends Controller
      */
     public function destroy(Activity $activity)
     {
+        $this->authorize('activities.delete', $activity);
+
         try {
             $activity->delete();
         } catch (Exception $e) {
